TheHeader3: Refactor around a new mobile design idea
I commented out a bunch of stuff to be able to revert back to
it in the future.
Added some div wrappers around stuff cause of flexbox.
Have the idea of using ::after and content= to change the
innerHTML of the <h2>s on mobile, ie: abbreviated to full from
mobile to desktop.
Thinking of going with no flexbox in the header, and use absolute
positioning on the button to then open up the menu nav, like in
this tutorial, https://www.w3schools.com/howto/howto_js_mobile_navbar.asp.